+++ crates/bevy_plugin/src/localization/strings_file.rs
@@ -2,6 +2,7 @@ use crate::prelude::*;
 use anyhow::bail;
 use bevy::prelude::*;
 use bevy::reflect::TypeUuid;
+use bevy::utils::HashMap;
 use seldom_fn_plugin::FnPluginExt;
 use sha2::{Digest, Sha256};
 
@@ -25,22 +26,27 @@ pub(crate) fn strings_file_plugin(app: &mut App) {
 #[non_exhaustive]
 pub(crate) struct StringsFile(pub(crate) Vec<StringsFileRecord>);
 
-impl StringsFile {
-    pub(crate) fn new_with_single_language(records: Vec<StringsFileRecord>) -> Result<Self> {
-        if let Some(language) = records.first().map(|record| &record.language) {
-            for record in records.iter().skip(1) {
-                if record.language != *language {
-                    bail!("Loaded strings file with mixed languages records must have the same language. Expected \"{}\", got \"{}\" in record: {:#?}",
-                    language,
-                    record.language,
-                    record)
-                }
-            }
-        }
-        Ok(Self(records))
-    }
-    pub(crate) fn language(&self) -> Option<&Language> {
-        self.0.first().map(|record| &record.language)
+/// Generates a string with the line metadata. This string is intended
+/// to be used in the "comment" column of a strings table CSV. Because
+/// of this, it will ignore the line ID if it exists (which is also
+/// part of the line metadata).
+///
+/// ## Return value
+/// A string prefixed with "Line metadata: ", followed by each
+/// piece of metadata separated by whitespace. If no metadata exists or
+/// only the line ID is part of the metadata, returns an empty string
+/// instead.
+fn read_comments(metadata: &[String]) -> String {
+    // Adapted from <https://github.com/YarnSpinnerTool/YarnSpinner-Unity/blob/462c735766a4c4881cd1ef1f15de28c83b2ba0a8/Editor/Importers/YarnProjectImporter.cs#L652>
+    let cleaned_metadata: Vec<_> = metadata
+        .iter()
+        .filter(|metadata| !metadata.starts_with("line:"))
+        .cloned()
+        .collect();
+    if cleaned_metadata.is_empty() {
+        String::new()
+    } else {
+        format!("Line metadata: {}", cleaned_metadata.join(" "))
     }
 }
 
@@ -107,5 +113,62 @@ impl LanguagesToStringsFiles {
         self.0
             .iter()
             .find_map(|(lang, h)| (h == handle).then_some(lang))
+    }
+}
+
+impl StringsFile {
+    pub(crate) fn new_with_single_language(records: Vec<StringsFileRecord>) -> Result<Self> {
+        if let Some(language) = records.first().map(|record| &record.language) {
+            for record in records.iter().skip(1) {
+                if record.language != *language {
+                    bail!("Loaded strings file with mixed languages records must have the same language. Expected \"{}\", got \"{}\" in record: {:#?}",
+                    language,
+                    record.language,
+                    record)
+                }
+            }
+        }
+        Ok(Self(records))
+    }
+    pub(crate) fn language(&self) -> Option<&Language> {
+        self.0.first().map(|record| &record.language)
+    }
+    pub(crate) fn from_yarn_files<'a>(
+        language: impl Into<Language>,
+        files: impl Iterator<Item = &'a YarnFile>,
+    ) -> Self {
+        let mut files: Vec<_> = files
+            .flat_map(|yarn_file| {
+                yarn_file
+                    .string_table
+                    .iter()
+                    .map(|(id, line_info)| (id, line_info, yarn_file.file.file_name.as_str()))
+                    .collect::<Vec<_>>()
+            })
+            .collect();
+        files.sort_by(
+            |(_, lhs_string_info, lhs_file_name), (_, rhs_string_info, rhs_file_name)| {
+                lhs_file_name.cmp(rhs_file_name).then(
+                    lhs_string_info
+                        .line_number
+                        .cmp(&rhs_string_info.line_number),
+                )
+            },
+        );
+        let language = language.into();
+        let strings_file_records = files
+            .into_iter()
+            .map(|(line_id, string_info, file_name)| StringsFileRecord {
+                language: language.clone(),
+                id: line_id.clone(),
+                text: string_info.text.clone(),
+                file: file_name.to_string(),
+                node: string_info.node_name.clone(),
+                line_number: string_info.line_number,
+                lock: Lock::compute_from(&string_info.text),
+                comment: read_comments(&string_info.metadata),
+            })
+            .collect();
+        Self(strings_file_records)
     }
 }
